requirements.txt
Flask
google-generativeai
requests
pandas
gunicorn
python-dotenv